# backend/docs/swagger.yaml
definitions:
  models.PollInfo:
    properties:
      description:
        type: string
      pollType:
        type: string
      title:
        type: string
    type: object
host: http://localhost:3001
info:
  contact: {}
  description: This is the API for the PartyPoll web application
  title: PartyPoll API
  version: "1.0"
paths:
  /login:
    post:
      responses: {}
      summary: Login an existing user
      tags:
      - Auth
  /polls:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PollInfo'
            type: array
      summary: Get all polls w/o results
      tags:
      - Polls
    post:
      parameters:
      - description: Add Poll
        in: body
        name: poll
        required: true
        schema:
          $ref: '#/definitions/models.PollInfo'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PollInfo'
            type: array
      summary: Post a poll
      tags:
      - Polls
  /polls/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Poll ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Delete a poll
      tags:
      - Polls
    get:
      parameters:
      - description: Poll ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get a poll and it's results
      tags:
      - Polls
    post:
      parameters:
      - description: Poll ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Post a poll result
      tags:
      - Polls
  /refresh-token:
    post:
      responses: {}
      summary: Create a refresh token
      tags:
      - Auth
swagger: "2.0"